Move curve visualization to elevated submodule
