+++ qiskit_experiments/curve_analysis/__init__.py
@@ -481,15 +481,6 @@ Data Classes
     ParameterRepr
     FitOptions
 
-Visualization
-=============
-
-.. autosummary::
-    :toctree: ../stubs/
-
-    BaseCurveDrawer
-    MplCurveDrawer
-
 Standard Analysis Library
 =========================
 
@@ -565,7 +556,6 @@ from .curve_fit import (
     process_curve_data,
     process_multi_curve_data,
 )
-from .visualization import BaseCurveDrawer, MplCurveDrawer
 from . import guess
 from . import fit_function
 from . import utils
@@ -580,6 +570,3 @@ from .standard_analysis import (
     ErrorAmplificationAnalysis,
     BlochTrajectoryAnalysis,
 )
-
-# deprecated
-from .visualization import plot_curve_fit, plot_errorbar, plot_scatter, FitResultPlotters
