+++ qiskit_experiments/curve_analysis/base_curve_analysis.py
@@ -24,7 +24,7 @@ from qiskit_experiments.data_processing import DataProcessor
 from qiskit_experiments.data_processing.processor_library import get_processor
 from qiskit_experiments.framework import BaseAnalysis, AnalysisResultData, Options, ExperimentData
 from .curve_data import CurveData, ParameterRepr, CurveFitResult
-from .visualization import MplCurveDrawer, BaseCurveDrawer
+from qiskit_experiments.visualization import MplCurveDrawer, BaseCurveDrawer
 
 PARAMS_ENTRY_PREFIX = "@Parameters_"
 DATA_ENTRY_PREFIX = "@Data_"
