+++ qiskit_experiments/curve_analysis/composite_curve_analysis.py
@@ -25,7 +25,7 @@ from qiskit_experiments.framework import BaseAnalysis, ExperimentData, AnalysisR
 from .base_curve_analysis import BaseCurveAnalysis, PARAMS_ENTRY_PREFIX
 from .curve_data import CurveFitResult
 from .utils import analysis_result_to_repr, eval_with_uncertainties
-from .visualization import MplCurveDrawer, BaseCurveDrawer
+from qiskit_experiments.visualization import MplCurveDrawer, BaseCurveDrawer
 
 
 class CompositeCurveAnalysis(BaseAnalysis):
