+++ qiskit_experiments/visualization/curves.py
